requirements.txt
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
glfw==2.7.0
idna==3.7
keyboard==0.13.5
mouse==0.7.1
MouseInfo==0.1.3
mss==9.0.1
mypy-extensions==1.0.0
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
Py-Keyboard==0.0.2
PyAutoGUI==0.9.54
PyDirectInput==1.0.4
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyOpenGL==3.1.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
pytweening==1.2.0
pywin32==306
requests==2.31.0
tomli==2.0.1
typing-extensions==4.11.0
urllib3==2.2.1
zero-hid==0.1.6
2captcha-python==1.2.8